+++ garden_app/index.ts
@@ -9,6 +9,9 @@ if (Platform.OS === 'android') {
   // Defines the background weather-alert task (it must exist before Android runs it).
   // eslint-disable-next-line @typescript-eslint/no-require-imports
   require('./src/services/alerts/backgroundAlerts');
+  // Defines the background "new version available" check.
+  // eslint-disable-next-line @typescript-eslint/no-require-imports
+  require('./src/services/updates/updateTask');
   // eslint-disable-next-line @typescript-eslint/no-require-imports
   const { registerWidgetTaskHandler } = require('react-native-android-widget');
   // eslint-disable-next-line @typescript-eslint/no-require-imports